Fixed test suites again
