+++ core/src/main/java/eu/tsystems/mms/tic/testframework/utils/ProxyUtils.java
@@ -1,7 +1,7 @@
 /*
  * Testerra
  *
- * (C) 2020, Mike Reiche, T-Systems Multimedia Solutions GmbH, Deutsche Telekom AG
+ * (C) 2021, Mike Reiche,  T-Systems Multimedia Solutions GmbH, Deutsche Telekom AG
  *
  * Deutsche Telekom AG and all other contributors /
  * copyright owners license this file to you under the Apache
@@ -17,10 +17,10 @@
  * KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations
  * under the License.
- *
  */
 package eu.tsystems.mms.tic.testframework.utils;
 
+import eu.tsystems.mms.tic.testframework.common.PropertyManager;
 import java.io.UnsupportedEncodingException;
 import java.net.MalformedURLException;
 import java.net.URL;
@@ -57,14 +57,14 @@ public class ProxyUtils {
 
         String proxyString = "";
 
-        final String proxyHost = System.getProperty(proxyType + ".proxyHost");
+        final String proxyHost = PropertyManager.getProperty(proxyType + ".proxyHost");
         if (proxyHost != null) {
             proxyString += proxyHost;
         } else {
             return null;
         }
 
-        final String proxyPort = System.getProperty(proxyType + ".proxyPort");
+        final String proxyPort = PropertyManager.getProperty(proxyType + ".proxyPort");
         if (proxyPort != null) {
             proxyString += ":" + proxyPort;
         }
@@ -86,12 +86,12 @@ public class ProxyUtils {
 
         try {
 
-            final String user = System.getProperty(proxyType + ".proxyUser");
+            final String user = PropertyManager.getProperty(proxyType + ".proxyUser");
             if (org.apache.commons.lang3.StringUtils.isNotEmpty(user)) {
                 proxyUrlString += URLEncoder.encode(user, urlEncoding);
             }
 
-            final String password = System.getProperty(proxyType + ".proxyPassword");
+            final String password = PropertyManager.getProperty(proxyType + ".proxyPassword");
             if (org.apache.commons.lang3.StringUtils.isNotEmpty(password)) {
                 proxyUrlString += ":" + URLEncoder.encode(password, urlEncoding);
             }
