+++ integration-tests/src/test/java/eu/tsystems/mms/tic/testframework/test/pagefactory/PageFactoryTest.java
@@ -23,6 +23,8 @@
 package eu.tsystems.mms.tic.testframework.test.pagefactory;
 
 import eu.tsystems.mms.tic.testframework.AbstractTestSitesTest;
+import eu.tsystems.mms.tic.testframework.common.PropertyManager;
+import eu.tsystems.mms.tic.testframework.constants.TesterraProperties;
 import eu.tsystems.mms.tic.testframework.core.pageobjects.testdata.BasePage;
 import eu.tsystems.mms.tic.testframework.core.pageobjects.testdata.BasePage2016;
 import eu.tsystems.mms.tic.testframework.core.pageobjects.testdata.PageWithExistingElement;
@@ -38,10 +40,13 @@ import eu.tsystems.mms.tic.testframework.core.testpage.TestPage;
 import eu.tsystems.mms.tic.testframework.execution.testng.AssertCollector;
 import eu.tsystems.mms.tic.testframework.pageobjects.Page;
 import eu.tsystems.mms.tic.testframework.pageobjects.factory.PageFactory;
+import eu.tsystems.mms.tic.testframework.report.Report;
+import eu.tsystems.mms.tic.testframework.report.TesterraListener;
 import eu.tsystems.mms.tic.testframework.webdrivermanager.AbstractWebDriverRequest;
 import eu.tsystems.mms.tic.testframework.webdrivermanager.UnspecificWebDriverRequest;
 import eu.tsystems.mms.tic.testframework.webdrivermanager.WebDriverManager;
 import eu.tsystems.mms.tic.testframework.webdrivermanager.WebDriverManagerConfig;
+import java.io.File;
 import java.net.MalformedURLException;
 import java.util.concurrent.atomic.AtomicBoolean;
 import org.openqa.selenium.WebDriver;
@@ -164,5 +169,36 @@ public class PageFactoryTest extends AbstractTestSitesTest {
 
         testPage.checkPage();
         Assert.assertTrue(atomicBoolean.get());
+    }
+
+    @Test
+    public void testT08_CheckPage_ScreenshotOnLoad() {
+
+        final File reportScreenshotDirectory = TesterraListener.getReport().getReportDirectory(Report.SCREENSHOTS_FOLDER_NAME);
+        Assert.assertNotNull(reportScreenshotDirectory);
+
+        final WebDriver driver = WebDriverManager.getWebDriver();
+
+        final int fileCountBeforeAction = getNumFiles(reportScreenshotDirectory);
+        PropertyManager.getTestLocalProperties().setProperty(TesterraProperties.SCREENSHOT_ON_PAGELOAD, "false");
+        new PageWithExistingElement(driver);
+
+        final int fileCountAfterCheckPageWithoutScreenshot = getNumFiles(reportScreenshotDirectory);
+        Assert.assertEquals(fileCountBeforeAction, fileCountAfterCheckPageWithoutScreenshot, "Record Screenshot count not altered.");
+
+        PropertyManager.getTestLocalProperties().setProperty(TesterraProperties.SCREENSHOT_ON_PAGELOAD, "true");
+        new PageWithExistingElement(driver);
+        final int fileCountAfterCheckPageWithScreenshot = getNumFiles(reportScreenshotDirectory);
+
+        Assert.assertNotEquals(fileCountAfterCheckPageWithoutScreenshot, fileCountAfterCheckPageWithScreenshot, "Record Screenshot count altered.");
+    }
+
+    private int getNumFiles(File directory) {
+        File[] files = directory.listFiles();
+        if (files == null) {
+            return 0;
+        } else {
+            return files.length;
+        }
     }
 }
