+++ integration-tests/src/test/resources/Integration.xml
@@ -1,60 +1,11 @@
 <!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
 <suite name="Testerra Integration Test Suite" verbose="10" configfailurepolicy="continue" thread-count="10" parallel="false">
-    <test name="Core" parallel="methods">
-        <groups>
-            <run>
-                <exclude name="SEQUENTIAL"/>
-            </run>
-        </groups>
-        <packages>
-            <package name="eu.tsystems.mms.tic.testframework.test.common"/>
-            <package name="eu.tsystems.mms.tic.testframework.test.reporting"/>
-            <package name="eu.tsystems.mms.tic.testframework.test.testdata"/>
-        </packages>
-    </test>
-
-    <test name="Sequential tests" parallel="false">
-        <groups>
-            <run>
-                <include name="SEQUENTIAL"/>
-            </run>
-        </groups>
-        <packages>
-            <package name="eu.tsystems.mms.tic.testframework.test.*"/>
-        </packages>
-    </test>
     <suite-files>
         <suite-file path="GuiElementStandard.xml"></suite-file>
         <suite-file path="GuiElementVariations.xml"></suite-file>
         <suite-file path="PageObjects.xml"></suite-file>
         <suite-file path="Localization.xml"></suite-file>
+        <suite-file path="Core.xml"></suite-file>
+        <suite-file path="Execution.xml"></suite-file>
     </suite-files>
-
-    <test name="Events" parallel="false">
-        <packages>
-            <package name="eu.tsystems.mms.tic.testframework.test.events"/>
-        </packages>
-    </test>
-
-    <test name="WebDriver" parallel="methods">
-        <groups>
-            <run>
-                <exclude name="SEQUENTIAL"/>
-            </run>
-        </groups>
-        <packages>
-            <package name="eu.tsystems.mms.tic.testframework.test.webdrivermanager"/>
-        </packages>
-    </test>
-
-    <test name="Utils" parallel="methods">
-        <groups>
-            <run>
-                <exclude name="SEQUENTIAL"/>
-            </run>
-        </groups>
-        <packages>
-            <package name="eu.tsystems.mms.tic.testframework.test.utils"/>
-        </packages>
-    </test>
 </suite>
